Minor modification to test functionality
Some of this code hasn't been used/tested with for several years,
and the SaC compiler has gone through a few changes. This commit
makes minor changes to make *most* of the benchmarks compile and
run. A few *do not* work, an possibly never did.

As of this commit:
  (WORKS) backprop
     (NO) bfs --------------- implementation incomplete
     (NO) cfd --------------- implementation incomplete
  (WORKS) hotspot
  (WORKS) kmeans
(PARTIAL) leukocyte
(PARTIAL) lud
  (WORKS) nw
(PARTIAL) particlefilter
  (WORKS) pathfinder
  (WORKS) srad